requirements.txt
behave==1.2.6
behave-django==1.1.0
dj-database-url==0.5.0
Django==2.0
django-countries==5.3.2
django-crispy-forms==1.7.2
django-debug-toolbar==1.10.1
django-extensions==2.1.3
django-image-cropping==1.2.0
django-js-asset==1.1.0
django-model-utils==3.1.2
django-money==0.14.3
django-mptt==0.9.1
django-phonenumber-field==2.1.0
django-trumbowyg==1.1.0
djangorestframework==3.9.0
easy-thumbnails==2.5
phonenumbers==8.10.0
Pillow==5.3.0
python-decouple==3.1
tox==3.5.3
